requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
passlib==1.7.4
Pillow==9.4.0
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.2
sqlparse==0.4.3
typing-extensions==4.5.0
